Rework core for efficiency
